AIP 4290-compliant Docker Interface for protofuzz.
Google's API Improvement Proposals describes a standard
Docker Interface for protoc plugins and their dependencies:
https://google.aip.dev/client-libraries/4290

This patch implements such an image and pushes it to
docker.io/hxtk/protofuzz.

TODO: Include googleapis protos

The googleapi protos are required by the AIP 4290 specification,
but have not been included in this initial image. Frankly, it's
a lot of work if I have to import them all manually into a TAR
and I am still searching for an automated way to do it.